# Optimise chunk loading