I started doing some benchmarking on the code that loads chunk since I noticed the game becomes a bit jumpy when entering a new area. After doing a bit of testing with a few different methods of optimisation, it seems like the main causes were:
- The chunk `DashMap` hash function
- The locking while loading a chunk
- The `BlockStorage` structure

The first thing I noticed was that a large amount of the time was used up by `DashMap` hash function. I tried to switch to a faster one, but nothing really seemed to beat the custom FNV hash that we have already. Since the chunk hashmap is indexed with two i32s, I tried to use a `BTreeMap` instead since the comparison should be faster than using a hash function for such a simple key. It ended up dropping from ~6.2ms to ~4.1ms per chunk.

One of the downsides of using `BTreeMap` rather than `DashMap` was that chunks couldn't be loaded concurrently since it keeps a write lock for the entire `load_chunk()` call. Using `DashMap` allows for locking multiple chunks if accessing values in different shards inside the `DashMap`. To resolve this issue while still using a `BTreeMap`, I changed `load_chunk()` to make a copy of the chunk at the start, update the chunk, and then write the copy back at the end. I was expecting this to add a little more time to the process, but I couldn't see any difference on the benchmark results. This also allows any number of chunks to be loaded concurrently, only blocking other threads while attempting to replace the chunk at the end.

The next thing I found was the `BlockStorage` struct. It seemed to be using a structure similar to the palette structure the chunk is sent in. When I switched this to use a `Vec<Block>`, it dropped from ~4.1ms to ~2.9ms per chunk. I tested the effect this had on memory by running the client with both implementations, and it seems like it only used ~20mb of extra memory.

After making these changes, the game was noticeably less jumpy when going into new regions. I think the majority of this was changing the locking structure, since anything that needed to read the world state was less likely to conflict with the `load_chunk()` write lock. The other optimisations are less important but still make chunks load faster in new regions.