@terrarier2111 2022-10-17
All in all, this change looks great (if the perf stats are correct) - although it leads to some extra work for me when working on rendering