@nathanruiz 2022-10-19
Which part makes the rendering code harder? I can always remove part of this, or apply them after the new renderer is merged in.