@terrarier2111 2022-10-20
> Which part makes the rendering code harder? I can always remove part of this, or apply them after the new renderer is merged in.

Its alr, i'll probably need a decent amount of time to come until i can finish the rendering update, i don't wanna block your amazing progress