@nathanruiz 2023-01-12
@terrarier2111 Yeah, I have no issues with this being merged. I think I've got some future ideas for optimisation, but they require more changes after the textures rework is complete.